# Fix: Optimize some pages load speed (metadata curator, summary, agents admin)