--- a/app/components/ontology_subscribe_button_component/ontology_subscribe_button_component.html.haml
+++ b/app/components/ontology_subscribe_button_component/ontology_subscribe_button_component.html.haml
@@ -1,4 +1,4 @@
-%div{@controller_params, style: 'margin-left: 10px;'}
+%div{@controller_params, style: 'margin-left: 10px;', id: @id}
   = render PillButtonComponent.new do
     %a.d-flex.align-items-center{href: @link}
       = inline_svg_tag "eye.svg"