--- a/app/controllers/ontologies_controller.rb
+++ b/app/controllers/ontologies_controller.rb
@@ -56,17 +56,17 @@ class OntologiesController < ApplicationController
     ] + update_filters_counts
 
     streams =if params[:page].nil?
-               [
-                 prepend('ontologies_list_container', partial: 'ontologies/browser/ontologies'),
-                 prepend('ontologies_list_container') {
-                   helpers.turbo_frame_tag("ontologies_filter_count_request") do
-                     helpers.browser_counter_loader
-                   end
-                 }
-               ]
-             else
-               [replace("ontologies_list_view-page-1", partial: 'ontologies/browser/ontologies')]
-             end
+                [
+                  prepend('ontologies_list_container', partial: 'ontologies/browser/ontologies'),
+                  prepend('ontologies_list_container') {
+                    helpers.turbo_frame_tag("ontologies_filter_count_request") do
+                      helpers.browser_counter_loader
+                    end
+                  }
+                ]
+              else
+                [replace("ontologies_list_view-page-1", partial: 'ontologies/browser/ontologies')]
+              end
 
     render turbo_stream: streams + count_streams
   end
@@ -246,7 +246,7 @@ class OntologiesController < ApplicationController
     # See: https://github.com/ncbo/bioportal_web_ui/issues/133.
     data_pages = KNOWN_PAGES - %w[summary notes]
     if @ontology.summaryOnly && params[:p].present? && data_pages.include?(params[:p].to_s)
-        redirect_to(ontology_path(params[:ontology]), status: :temporary_redirect) and return
+      redirect_to(ontology_path(params[:ontology]), status: :temporary_redirect) and return
     end
 
     #@ob_instructions = helpers.ontolobridge_instructions_template(@ontology)
@@ -346,6 +346,28 @@ class OntologiesController < ApplicationController
     end
   end
 
+  def subscriptions
+    ontology_id = params[:ontology_id]
+    return not_found if ontology_id.nil?
+
+    ontology_acronym = ontology_id.split('/').last
+
+    if session[:user].nil?
+      link = "/login?redirect=#{request.url}"
+      subscribed = false
+      user_id = nil
+    else
+      user = LinkedData::Client::Models::User.find(session[:user].id).first
+      subscribed = helpers.subscribed_to_ontology?(ontology_acronym, user)
+      link = "javascript:void(0);"
+      user_id = user.id
+    end
+
+    count = helpers.count_subscriptions(ontology_id)
+    render inline:  helpers.turbo_frame_tag('subscribe_button') {
+      render_to_string(OntologySubscribeButtonComponent.new(id: 'subscribe_button', ontology_id: ontology_id, subscribed: subscribed, user_id: user_id, count: count, link: link),  layout: nil)
+    }
+  end
 
   def virtual
     redirect_new_api
@@ -362,7 +384,7 @@ class OntologiesController < ApplicationController
       render partial: 'ontologies/sections/widgets', layout: 'ontology_viewer'
     end
   end
-  
+
 
   def show_additional_metadata
     @metadata = submission_metadata