--- a/app/controllers/ontologies_controller.rb
+++ b/app/controllers/ontologies_controller.rb
@@ -461,12 +461,11 @@ class OntologiesController < ApplicationController
 
         target_id = relation_value
         target_in_portal = false
+        target_ont = nil
         # if we find our portal URL in the ontology URL, then we just keep the ACRONYM to try to get the ontology.
-        relation_value = relation_value.split('/').last if relation_value.include?($UI_URL)
-
-        # Use acronym to get ontology from the portal
-        target_ont = LinkedData::Client::Models::Ontology.find_by_acronym(relation_value).first
-        if target_ont
+        if relation_value.include?(portal_name)
+          relation_value = relation_value.split('/').last
+          target_ont = LinkedData::Client::Models::Ontology.find_by_acronym(relation_value).first
           target_id = target_ont.acronym
           target_in_portal = true
         end