--- a/app/controllers/ontologies_metadata_curator_controller.rb
+++ b/app/controllers/ontologies_metadata_curator_controller.rb
@@ -5,29 +5,16 @@ class OntologiesMetadataCuratorController < ApplicationController
   before_action :submission_metadata, only: [:result, :edit, :update, :show_metadata_by_ontology]
 
   def result
-    @ontologies_ids = params[:ontology] ? params[:ontology][:ontologyId] : []
+    @ontologies_ids = params[:ontology] ? Array(params[:ontology][:ontologyId]) : []
     @metadata_sel = params[:search] ? params[:search][:metadata] : []
     @show_submissions = !params[:show_submissions].nil?
-    @ontologies = []
     @submissions = []
 
-    if @ontologies_ids.nil? || @ontologies_ids.empty?
-      @ontologies = LinkedData::Client::Models::Ontology.all
-    else
-      @ontologies_ids.each do |data|
-        @ontologies << LinkedData::Client::Models::Ontology.find_by_acronym(data).first
-      end
-    end
-
     display_attribute = equivalent_properties(@metadata_sel) + %w[submissionId]
-    @ontologies.each do |ont|
-      if @show_submissions
-        submissions = ont.explore.submissions({ include: display_attribute.join(',') })
-      else
-        submissions = [ont.explore.latest_submission({ include: display_attribute.join(',') })]
-      end
-      submissions.each { |sub| append_submission(ont, sub) }
-    end
+
+    @submissions = LinkedData::Client::Models::OntologySubmission.all(acronym: @ontologies_ids.join('|'), display_links: false, display_context: false, include: display_attribute.join(','), include_status: 'RDF')
+    @submissions.reject!{|x| !@ontologies_ids.include?(x.id.split('/')[-3])} unless @ontologies_ids.empty?
+    @submissions.sort_by!{|x| x.id}
 
     respond_to do |format|
       format.html { redirect_to admin_index_path }