--- a/app/views/layouts/_ontology_viewer.html.haml
+++ b/app/views/layouts/_ontology_viewer.html.haml
@@ -56,7 +56,7 @@
                      selected: selected_section?(section_title),
                      page_name: ontology_viewer_page_name(@ontology.acronym, @concept&.prefLabel || '', section_title))
             - t.pinned_right do
-              %div{style: "visibility: #{ontology_data_section? ? 'visible' : 'hidden'}; border: none; outline: none; padding: 1px",
+              %div{style: "display: #{ontology_data_section? ? 'block' : 'none'}; border: none; outline: none; padding: 1px",
                 data: {'ontology-viewer-tabs-target': 'languageSelector' }}
                 = language_selector_tag(:content_language)
             - t.item_content do