--- a/app/views/submissions/edit.html.haml
+++ b/app/views/submissions/edit.html.haml
@@ -1,7 +1,7 @@
 - section = params[:section] || 'general'
 .center
   .edit-ontology-container
-    = turbo_frame_tag(params[:container_id]) do
+    = turbo_frame_tag(params[:container_id] || 'test') do
       = form_for :submission, url: ontology_submission_path(@ontology.acronym, params["id"]), html: { id: "ontology_submission_form", method: :put, multipart: true, 'data-turbo': true, 'data-turbo-frame': '_top', novalidate: 'true'} do
         .edit-ontology-title
           %div Edit ontology