### Prerequisites
* https://github.com/ontoportal-lirmm/ontologies_api_ruby_client/pull/14 

### Changes
* Optimize the summary page loading time ( 51.06% less time), by 
   * [update the summary watch button to be lazy loaded](https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/411/commits/0b65c2497d6e251e8aa6f1682d3a191a542883c6)
   * [optimize the submissions partial to not load the latest submission twice](https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/411/commits/58f7faceeccb48f4d714738a619464a38c4b1c11)
* Optimize submission update page loading time  (79.99% less time), by 
  * [optimize the submission update page speed](https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/411/commits/3e875371837a8900623e8b5840dbe9602d09bed7)
* Optimize metadata curator page loading time (97.1% less time), by 
  *  [optimize the metadata curator call to use the new submissions endpoint](https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/411/commits/69553afa8458bf8a378337211116b212b58bd08c)
* Optimize their agent admin page loading time (91.8% less time), by 
  * https://github.com/ontoportal-lirmm/ontologies_linked_data/commit/e128d4422ece20c62acc3c2032e3029f4f6d7719 

See detailed benchmark in the comment below