@syphax-bouazzouni 2023-12-18
### Agents  ✅
- Before
 ```ruby
            ## UI 
            irb(main):026> Benchmark.realtime {  LinkedData::Client::Models::Agent.all({invalidate_cache: true})}
            Getting: https://data.agroportal.lirmm.fr with {}
            Getting: https://data.agroportal.lirmm.fr/agents with {}
            => 1.6159074329771101
            irb(main):027> Benchmark.realtime {  LinkedData::Client::Models::Agent.all({invalidate_cache: true, include: 'all'})}
            Getting: https://data.agroportal.lirmm.fr with {}
            Getting: https://data.agroportal.lirmm.fr/agents with {:include=>"all"}
            => 31.894347736990312
            
            ## API 
            [22] pry(main)> Benchmark.realtime { LinkedData::Models::Agent.where.include(LinkedData::Models::Agent.goo_attrs_to_load([:all])).to_a } 
            => 0.3907707901671529
            [23] pry(main)> Benchmark.realtime { LinkedData::Models::Agent.load_agents_usages(LinkedData::Models::Agent.where.include(LinkedData::Models::Agent.goo_attrs_to_load([:all])).to_a) }
            => 18.57385656516999
```
            
- After  https://github.com/ontoportal-lirmm/ontologies_linked_data/commit/e128d4422ece20c62acc3c2032e3029f4f6d7719
            
```ruby
            ### UI 
            irb(main):030> Benchmark.realtime {  LinkedData::Client::Models::Agent.all({invalidate_cache: true})}
            Getting: https://data.agroportal.lirmm.fr with {}
            Getting: https://data.agroportal.lirmm.fr/agents with {}
            => 2.1963163029868156
            irb(main):031> Benchmark.realtime {  LinkedData::Client::Models::Agent.all({invalidate_cache: true, include: 'all'})}
            Getting: https://data.agroportal.lirmm.fr with {}
            Getting: https://data.agroportal.lirmm.fr/agents with {:include=>"all"}
            => 2.6242307079955935
            
            ### API
            [2] pry(main)> Benchmark.realtime { LinkedData::Models::Agent.where.include(LinkedData::Models::Agent.goo_attrs_to_load([:all])).to_a } 
            => 0.37851628847420216
            [3] pry(main)> Benchmark.realtime { LinkedData::Models::Agent.load_agents_usages(LinkedData::Models::Agent.where.include(LinkedData::Models::Agent.goo_attrs_to_load([:all])).to_a) }
            => 0.8850884884595871
```
            
            
 ### Summary page  ✅
- Before (31 request to API, total response time:  27.287156736002845s)
            
```ruby
            Started GET "/ontologies/WHEATPHENOTYPE" 
            
            Getting: https://data.agroportal.lirmm.fr/slices with {} (0.8719426310271956s)
            Getting: https://data.agroportal.lirmm.fr with {} (0.5007915890018921s)
            Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all", :include_views=>true} (1.6749098369909916s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include=>"all"} (0.8536582970118616s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/metrics with {} (0.6196307270147372s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/projects with {} (0.5343920940067619s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/analytics with {} (0.5147110710095149s)
            FAIRness service issue unreachable
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/views with {:include=>"all"} (0.358907472982537s)
            Getting: https://data.agroportal.lirmm.fr/categories with {} (0.41279732901602983s)
            
            Getting: https://data.agroportal.lirmm.fr/groups with {} (0.36467222901410423s)
            Getting: https://data.agroportal.lirmm.fr/groups with {} (0.3739724579791073s)
            Getting: https://data.agroportal.lirmm.fr/groups with {} (0.4971552220231388s)
            
            Getting: https://data.agroportal.lirmm.fr with {} (1.4704590500041377s)
            Getting: https://data.agroportal.lirmm.fr/users with {:include=>"all"} (1.6560365210170858s)
            Getting: https://data.agroportal.lirmm.fr/users with {:include=>"subscription", :display_context=>false, :display_links=>false} (1.4961505329993088s)
            
            Started GET "/ontologies/WHEATPHENOTYPE/submissions" 
            
            Getting: https://data.agroportal.lirmm.fr/slices with {} (1.8425917400163598s)
            Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all", :include_views=>true} (1.6591912480071187s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/submissions with {:include=>"submissionId,creationDate,released,modificationDate,submissionStatus,hasOntologyLanguage,version,diffFilePath,ontology"} (0.8058881250035483s)
             
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :display=>"submissionId", :include=>"all"} (0.42191318501136266s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (1.1579404840013012s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (0.46696320100454614s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (0.6056716369930655s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (0.6060693030012771s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (0.612728006002726s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (1.5124448410060722s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (0.5820022370025981s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (1.4374711389828008s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (0.5724861980124842s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (0.595302603993332s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include_status=>"ready", :include=>"all"} (0.5287278570176568s)
            ```
            
- After  (18 requests to the API, total response time:13.352862464000469s, 51.06% less)
            
            ```ruby
            Started GET "/ontologies/WHEATPHENOTYPE" 
            Getting: https://data.agroportal.lirmm.fr/slices with {} (0.8188329540134873s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE with {:include_views=>true} (1.6679701330140233s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/latest_submission with {:include=>"all"} (0.7342191549832933s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/metrics with {} (0.5149527960165869s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/projects with {} (0.5129906750225928s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/analytics with {} (0.4231351870112121s)
            FAIRness service issue unreachable
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/views with {:include=>"all"} (1.021268181008054s)
            Getting: https://data.agroportal.lirmm.fr/categories/TRAITS with {} (1.4589310270093847s)
            
            Getting: https://data.agroportal.lirmm.fr/groups/WHEAT with {} (0.46195283898850903s)
            Getting: https://data.agroportal.lirmm.fr/groups/D2KAB with {} (0.3907403089979198s)
            Getting: https://data.agroportal.lirmm.fr/groups/INRAE with {} (0.38944203799474053s)
            
            Started GET "/ontologies/WHEATPHENOTYPE/submissions" 
            Getting: https://data.agroportal.lirmm.fr/slices with {} (0.5228270990191959s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE with {} (0.42837247301940806s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/submissions with {:include=>"submissionId,creationDate,released,modificationDate,submissionStatus,hasOntologyLanguage,version,diffFilePath,ontology"} (1.4856416810071096s)
            
            Started GET "/ontologies/WHEATPHENOTYPE/subscriptions" 
            Getting: https://data.agroportal.lirmm.fr/slices with {} (0.3582408800139092s)
            Getting: https://data.agroportal.lirmm.fr/users/syphax+bouazzouni with {} (0.4949605920119211s)
            Getting: https://data.agroportal.lirmm.fr/users with {:include=>"subscription", :display_context=>false, :display_links=>false} (0.5689596759912092s)
```
 ### Save metadata  ✅
- Before (15 requests to the API, Total Response Time ≈ 13.3257690559891s)
        
```ruby
        Started PUT "/ontologies/WHEATPHENOTYPE/submissions/11" for 192.168.65.1 at 2023-12-17 03:36:17 +0000
        
        Getting: https://data.agroportal.lirmm.fr/slices with {} (0.43723661999683827s)
        start update
        start update ontology
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE with {} (0.2069408730021678s)
        invalidate cache all
        Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all", :include_views=>true} (1.135303320013918s)
        Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all"} (1.0787174180150032s)
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE with {} (0.3540415579918772s)
        
        start update submission
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/submissions/11 with {:display=>"all", :include=>"all"} (0.48975725600030273s)
        update ontology summary only
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE with {:include=>"all"} (0.2820891629962716s)
        invalidate cache all
        Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all"} (1.0174597069853917s)
        Getting: https://data.agroportal.lirmm.fr with {} (0.3596201420004945s)
        Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all", :include_views=>true} (1.7241795949812513s)
        Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all"} (0.9491338549996726s)
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE with {} (0.29278956400230527s)
        invalidate cache all
        Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all"} (1.1268853299843613s)
        Getting: https://data.agroportal.lirmm.fr with {} (1.4538259789987933s)
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/submissions/11 with {} (1.754813955980353s)
        
        Redirected to http://localhost:3000/ontologies/WHEATPHENOTYPE
        end update
```
        
- After (5 requests to the API, Total Response Time ≈ 2.675883663945365s,  79.99% less)
        
```ruby
        Started PUT "/ontologies/WHEATPHENOTYPE/submissions/11" for 192.168.65.1 at 2023-12-17 04:18:58 +0000
        
        Getting: https://data.agroportal.lirmm.fr/slices with {} (0.2552103519847151s)
        start update
        start update ontology
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE with {} (0.43316036998294294s)
        invalidate cache all
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE with {} (0.4646476349735167s)
        start update submission
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/submissions/11 with {:include=>"all"} (0.4673299429996405s)
        invalidate cache all
        Getting: https://data.agroportal.lirmm.fr/ontologies/WHEATPHENOTYPE/submissions/11 with {} (1.0545353640045505s)
        Redirected to http://localhost:3000/ontologies/WHEATPHENOTYPE
        end update
```
        
  ### Curator ✅
- before (143 requests to API,  total response time: 73.59983067121386s)
            
```ruby
            
            Getting: https://data.agroportal.lirmm.fr/slices with {} (0.4259469490207266s)
            Getting: https://data.agroportal.lirmm.fr/ontologies with {:include=>"all"} (1.3401180359942373s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/E-PHY/latest_submission with {:include=>"abstract,submissionId"} (0.43087202799506485s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_330/latest_submission with {:include=>"abstract,submissionId"} (0.361107604985591s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/AGRO/latest_submission with {:include=>"abstract,submissionId"} (0.35653764600283466s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_341/latest_submission with {:include=>"abstract,submissionId"} (0.33050059899687767s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/MEAT-T/latest_submission with {:include=>"abstract,submissionId"} (0.36843643002794124s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/SENSORML/latest_submission with {:include=>"abstract,submissionId"} (0.3464520720008295s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/SIO/latest_submission with {:include=>"abstract,submissionId"} (0.33479020700906403s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_320/latest_submission with {:include=>"abstract,submissionId"} (0.3400632810080424s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/EBO/latest_submission with {:include=>"abstract,submissionId"} (0.31141855398891494s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/RO/latest_submission with {:include=>"abstract,submissionId"} (0.2840256519848481s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/GECO/latest_submission with {:include=>"abstract,submissionId"} (1.4783019199967384s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_331/latest_submission with {:include=>"abstract,submissionId"} (0.33597830799408257s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/SDNDEV-TEST/latest_submission with {:include=>"abstract,submissionId"} (0.3005179439787753s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/INIA-RDM/latest_submission with {:include=>"abstract,submissionId"} (0.4669017209962476s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/VBO/latest_submission with {:include=>"abstract,submissionId"} (0.31066214400925674s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/SDGIO/latest_submission with {:include=>"abstract,submissionId"} (0.3623409619904123s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_321/latest_submission with {:include=>"abstract,submissionId"} (0.3989575729938224s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/SIREN/latest_submission with {:include=>"abstract,submissionId"} (0.25879321197862737s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/FALDO/latest_submission with {:include=>"abstract,submissionId"} (0.4047464120085351s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/EFO/latest_submission with {:include=>"abstract,submissionId"} (1.4349618909764104s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/VARIO/latest_submission with {:include=>"abstract,submissionId"} (0.2781464940053411s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/SSN/latest_submission with {:include=>"abstract,submissionId"} (0.36977093800669536s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CVO/latest_submission with {:include=>"abstract,submissionId"} (0.333509464020608s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/AGRORDF/latest_submission with {:include=>"abstract,submissionId"} (0.3190122879750561s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/WCACROPS/latest_submission with {:include=>"abstract,submissionId"} (0.2704014180053491s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/OBOE/latest_submission with {:include=>"abstract,submissionId"} (0.452480217005359s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_365/latest_submission with {:include=>"abstract,submissionId"} (0.34717402601381764s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/BIOREFINERY/latest_submission with {:include=>"abstract,submissionId"} (0.37172165798256174s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CPC/latest_submission with {:include=>"abstract,submissionId"} (0.42183868400752544s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_322/latest_submission with {:include=>"abstract,submissionId"} (0.2944680769869592s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/FOBI/latest_submission with {:include=>"abstract,submissionId"} (0.33096307300729677s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CDAO/latest_submission with {:include=>"abstract,submissionId"} (0.3728140230232384s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_325/latest_submission with {:include=>"abstract,submissionId"} (0.31974423900828697s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/PIZZA/latest_submission with {:include=>"abstract,submissionId"} (0.35532932201749645s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/AQFO/latest_submission with {:include=>"abstract,submissionId"} (0.3953713570081163s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/PR/latest_submission with {:include=>"abstract,submissionId"} (0.2717428150062915s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/SFWO/latest_submission with {:include=>"abstract,submissionId"} (0.28472482098732144s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_345/latest_submission with {:include=>"abstract,submissionId"} (1.41935532700154s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/ATC/latest_submission with {:include=>"abstract,submissionId"} (0.2906604559975676s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/CO_125/latest_submission with {:include=>"abstract,submissionId"} (0.36497895198408514s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/TO/latest_submission with {:include=>"abstract,submissionId"} (0.45950186398113146s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/TIME/latest_submission with {:include=>"abstract,submissionId"} (0.3139042299881112s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/GENO/latest_submission with {:include=>"abstract,submissionId"} (0.3936639309977181s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/PO2/latest_submission with {:include=>"abstract,submissionId"} (0.41684210899984464s)
            Getting: https://data.agroportal.lirmm.fr/ontologies/BCO/latest_submission with {:include=>"abstract,submissionId"} (0.8354450919723604s)
…